eliminated all of the warnings #306
Many were due to number formatting in strings while others were related
to pre-declaring methods which does not have to be done in modern
Objective-C and Xcode 5. I also removed ACEAutocomplete which does not
appear to be used in the app and changed MBProgressHUD to be used as a
CocoaPod reference. Now when warnings are discovered they can be
address immediately so they do not add up.