Fix for #394
I updated the common section header which does include these space with
a gradient on top but I docked everything else to the bottom so I could
set the height of the header section to 47 which hides the header. Now
it looks as intended and could be done for any section header by just
changing the height.