implemented highest rated for a spot in the pullup UI
@MariusVolkhart @claudobahn

You can see in this commit that the ID values for the slider template
for beer, wine and cocktail emphasis as well as the drink types are now
hard-coded as constants. Since these are primary keys in the database
for an auto-incremented field this could be a potential problem. Since
Production is duplicated to Staging these ID values are identical
across databases, but going forward these values must also be enforced
on the database. And future ID values which may need to be referenced
in a similar way may need to be set intentionally.